Fix manual check of parameters to skip
The `p` constant originally held an array of two values representing gompertz shape values.

This was split into two separate parameters and renamed from `p` to `gompertz_p1` and `gompertz_p2` for clarity.

Neglected to update the single scenario script to reflect this change which has led the the issues reported by @KRN01 and @Rosejoycrocker